requirements.txt
flask
flask-sqlalchemy
flask_limiter
requests_oauthlib
Flask-SSLify
flask_socketio
paypalrestsdk
git+https://github.com/TitanEmbeds/discord.py@rewrite#egg=discord.py
asyncio_extras
kombu
redis
aioredis
Flask-Babel
patreon
flask-redis
gino
sqlalchemy==1.2.10
asyncio_redis
raven[flask]
eventlet==0.24.1
python-dotenv